+++ src/AlertManagerTG.py
@@ -27,8 +27,7 @@ def exit_gracefully(signum, frame):
     signal.signal(signal.SIGINT, exit_gracefully)
 
 def main():
-    logging.basicConfig(format="%(asctime)s %(levelname)s\t%(name)s:%(message)s", level= _config._settings['global']['log_level'])
-    logging.getLogger("server.worker").setLevel(_config.logLevel)
+    logging.basicConfig(format="%(asctime)s %(levelname)s\t%(name)s:%(message)s", level=_config.logLevel)
     app = make_app()
     app.listen(_config.webport)
     try:
